Home page added (root)

+++ authentication/views.py
@@ -86,6 +86,12 @@ class LoginAPIView(APIView):
 
 class LogoutAPIView(APIView):
     def get(self, request):
-        request.user.auth_token.delete()
-        logout(request)
-        return Response({"message": "User logout successfully!"})
+        if request.user.is_anonymous:
+            return Response({"error": "User is not authenticated!"}, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            try:
+                request.user.auth_token.delete()
+            except User.auth_token.RelatedObjectDoesNotExist:
+                return Response({"error": "User has no auth_token!"}, status=status.HTTP_400_BAD_REQUEST)
+            logout(request)
+            return Response({"message": "User logout successfully!"})
